requirements.txt
alembic==1.11.2
bcrypt==4.0.1
blinker==1.6.2
build==0.10.0
click==8.1.6
Flask==2.3.2
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pip-tools==7.3.0
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.19
tomli==2.0.1
typing_extensions==4.7.1
Werkzeug==2.3.6
zipp==3.16.2
